:truck: refactor: group modules to one directory (#13)
