+++ src/app.module.ts
@@ -1,12 +1,26 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserModule } from './domain/user/user.module';
-import { FamilyModule } from './domain/family/family.module';
+import {
+  FamilyModule,
+  UserModule,
+  FamilyMemberModule,
+  FamilyScheduleModule,
+  PhotoModule,
+  PostModule,
+} from './module';
 import { MysqlModule } from './infra/database/mysql.module';
 
 @Module({
-  imports: [UserModule, FamilyModule, MysqlModule],
+  imports: [
+    UserModule,
+    FamilyModule,
+    MysqlModule,
+    FamilyMemberModule,
+    FamilyScheduleModule,
+    PhotoModule,
+    PostModule,
+  ],
   controllers: [AppController],
   providers: [AppService],
 })
