+++ src/domain/post/index.ts
@@ -1,6 +1,3 @@
-export * from './dto/request/create-post.dto';
-export * from './dto/request/update-post.dto';
-export * from './dto/response/response-post.dto';
-export * from './post.module';
+export * from './dto';
 export * from './post.service';
 export * from './post.controller';
