+++ src/domain/user/dto/index.ts
@@ -1,2 +1,3 @@
 export * from './request/create-user.dto';
 export * from './request/update-user.dto';
+export * from './response/response-user-dto';
