+++ src/domain/user/dto/response/response-user-dto.ts
@@ -1,0 +1,3 @@
+export class ResponseUserDto{
+
+}
