+++ src/domain/user/index.ts
@@ -1,0 +1,3 @@
+export * from './dto';
+export * from './user.service';
+export * from './user.controller';
