+++ src/module/index.ts
@@ -1,0 +1,6 @@
+export * from './family-member.module';
+export * from './family.module';
+export * from './family-schedule.module';
+export * from './photo.module';
+export * from './user.module';
+export * from './post.module';
