+++ src/module/photo.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { PhotoController } from './photo.controller';
-import { PhotoService } from './photo.service';
+import { PhotoController } from '../domain/photo/photo.controller';
+import { PhotoService } from '../domain/photo/photo.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Photo } from '../../infra/entities';
+import { Photo } from '../infra/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Photo])],
