+++ src/module/post.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { PostController } from './post.controller';
-import { PostService } from './post.service';
+import { PostController } from '../domain/post/post.controller';
+import { PostService } from '../domain/post/post.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Post } from '../../infra/entities';
+import { Post } from '../infra/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Post])],
